requirements.txt
PySide6>=6.5
vgamepad>=0.1.0
pyserial>=3.5
pyusb>=1.3.0
libusb>=1.0.29
